Make ui more mobile friendly, unbloat Navigation component
